requirements.txt
pyzmq==19.0.2
redis==3.5.3